refactor: Move ApkSignerUtils to pro.sketchware package
