+++ app/src/main/java/pro/sketchware/utility/apk/ApkSignerUtils.java
@@ -1,4 +1,4 @@
-package mod.yamenher;
+package pro.sketchware.utility.apk;
 
 import android.os.Handler;
 import android.os.HandlerThread;
